Fixes more of the app to SvelteKit 5.0. Firebase integration still not working
