+++ client/src/routes/todos/+page.server.ts
@@ -1,0 +1,64 @@
+import {api} from './_api';
+
+export async function load({locals}) {
+  // locals.userid comes from src/hooks.js
+  const response = await api('GET', `todos/${locals.userid}`);
+
+  if (response.status === 404) {
+    // user hasn't created a todo list.
+    // start with an empty array
+    return {
+      body: {
+        todos: [],
+      }
+    };
+  }
+
+  if (response.status === 200) {
+    return {
+      todos: await response.json(),
+    };
+  }
+
+  return {
+    status: response.status,
+  };
+};
+
+export const actions = {
+  post: async ({request, locals}) => {
+    const form = await request.formData();
+  
+    await api('POST', `todos/${locals.userid}`, {
+      text: form.get('text'),
+    });
+  
+    return {};
+  },
+  patch: async ({request, locals}) => {
+    const form = await request.formData();
+  
+    await api('PATCH', `todos/${locals.userid}/${form.get('uid')}`, {
+      text: form.has('text') ? form.get('text') : undefined,
+      done: form.has('done') ? !!form.get('done') : undefined
+    });
+  
+    return redirect;
+  },
+  delete: async ({request, locals}) => {
+    const form = await request.formData();
+  
+    await api('DELETE', `todos/${locals.userid}/${form.get('uid')}`);
+  
+    return redirect;
+  },
+}
+
+// If the user has JavaScript disabled, the URL will change to
+// include the method override unless we redirect back to /todos
+const redirect = {
+  status: 303,
+  headers: {
+    location: '/todos',
+  }
+};
